refactor: onboarding api 파일 step 번호 제거
